### 2022-07-03 #4541 @Kevin-Mattheus-Moerman
@JorenSix thanks for this submission to JOSS. I will be looking for an editor for this submission. In the mean time can you address those potentially missing DOIs? :point_up: Once you've checked them  you can run `@editorialbot generate pdf` to update the draft, and you can call `@editorialbot check refereces` to check those DOIs again. Thanks
