+++ paper.bib
@@ -1,15 +1,3 @@
-@article{2001lemanipem-archive,
-  author       = {{Leman, Marc and Dierickx, Jelle and Martens, Gaëtan}},
-  issn         = {{0929-8215}},
-  journal      = {{JOURNAL OF NEW MUSIC RESEARCH}},
-  language     = {{eng}},
-  number       = {{4}},
-  pages        = {{389--393}},
-  title        = {{The IPEM-archive conservation and digitalization project.}},
-  volume       = {{30}},
-  year         = {{2001}},
-}
-
 
 @inproceedings{six2012ethnic_fingerprinter,
   author = {Joren Six and Olmo Cornelis},
@@ -47,7 +35,8 @@
   booktitle={Multimedia Signal Processing, 2002 IEEE Workshop on},
   pages={165--168},
   year={2002},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/MMSP.2002.1203273}
 }
 
 @inproceedings{Wang2003a,
@@ -75,7 +64,8 @@
   number={4},
   pages={775--785},
   year={2012},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi={10.1109/TASL.2012.2234114}
 }
 
 @InProceedings{six2018dupapps,
@@ -88,7 +78,8 @@
   publisher="Springer International Publishing",
   address="Cham",
   pages="101--113",
-  isbn="978-3-319-73165-0"
+  isbn="978-3-319-73165-0",
+  doi="10.1007/978-3-319-73165-0_10"
 }
 
 
@@ -161,17 +152,6 @@
   publisher={Acoustical Society of America}
 }
 
-@article{gabor1946theory,
-  title={Theory of communication. Part 1: The analysis of information},
-  author={Gabor, Dennis},
-  journal={Journal of the Institution of Electrical Engineers-Part III: Radio and Communication Engineering},
-  volume={93},
-  number={26},
-  pages={429--441},
-  year={1946},
-  publisher={IET}
-}
-
 @article{comer1979ubiquitous,
   title={Ubiquitous B-tree},
   author={Comer, Douglas},
@@ -220,17 +200,6 @@ year	= {2020},
 booktitle	= {ISMIR 2020 - 21st International Society for Music Information Retrieval Conference}
 }
 
-@article{marolt2019automatic,
-  title={Automatic Segmentation of Ethnomusicological Field Recordings},
-  author={Marolt, Matija and Bohak, Ciril and Kav{\v{c}}i{\v{c}}, Alenka and Pesek, Matev{\v{z}}},
-  journal={Applied Sciences},
-  volume={9},
-  number={3},
-  pages={439},
-  year={2019},
-  publisher={Multidisciplinary Digital Publishing Institute}
-}
-
 
 @inproceedings{sonnleitner2014quad_based_fingerprinter,
   author    = {Reinhard Sonnleitner and Gerhard Widmer},
@@ -267,21 +236,6 @@ booktitle	= {ISMIR 2020 - 21st International Society for Music Information Retri
   number={US7627477 B}
 }
 
-@article{Six2015,
-author="Six, Joren
-and Leman, Marc",
-title="Synchronizing multimodal recordings using audio-to-audio alignment",
-journal="Journal on Multimodal User Interfaces",
-year="2015",
-month="Sep",
-day="01",
-volume="9",
-number="3",
-pages="223--229",
-issn="1783-8738",
-doi="10.1007/s12193-015-0196-1",
-}
-
 @inproceedings{sonnleitner2016landmark,
   title={Landmark-based audio fingerprinting for DJ mix monitoring.},
[…]
