### 2022-07-07 #4554 @editorialbot
[REVIEW]: Panako: a scalable audio search system
**Submitting author:** @JorenSix (<a href="http://orcid.org/0000-0001-7671-1907">Joren Six</a>)
**Repository:** https://github.com/JorenSix/Panako
**Branch with paper.md** (empty if default branch):
**Version:** 2.1.0
