### 2022-07-08 #4554 @dpwe
* Although the readme invites people to contribute to the project, there are no specific instructions on contributing.  I also didn't see any specific instructions on how to report issues, but I think using the GitHub issue tracker is implicit.
* The documentation claims the uniqueness of the functionality is that it's an available, open-source, large-scale audio fingerprinter *which tolerates time scaling*.  The work acknowledges audfprint as an available open-source audio fingerprint engine, but that engine does not use a time-scaling-invariant algorithm.  However, I note that the Matlab version of audfprint addresses time-scaling through -warpmax and -warpstep options, which are discussed under "Sensitivity to Playback Speed" in https://www.ee.columbia.edu/~dpwe/resources/matlab/audfprint/ .  This is a different, pragmatic approach to handling the kinds of small time scaling that can be encountered in real signals, by simply re-running multiple versions of the query exhaustively spanning the range of tolerable time scalings.  In some cases, this reduces the risk of false positives compared to full time-scale invariance (at the cost of query processing time).  However, this functionality in audfprint is obscure (and inaccessible since that code requires Matlab).
* I didn't see any mention of the Python port https://github.com/dpwe/audfprint which at least escapes the Matlab dependency (but does not support warping queries IIRC). It's possible the author is unaware of it since it has not been widely publicized.  Unlike Panako, it is not actively supported.
* I don't think Panako needs to rely on its time-scale-invariant feature to justify its existence; instead, it is distinguished by being a well-documented, actively-supported open-source project.  However, it might be interesting to understand its performance relative to other algorithms.  Fingerprinting systems are mostly differentiated by their processing time/storage requirements, minimum query length, and SNR tolerance.  The ISMIR2014 paper looks at true positive rate as a function of time scaling, but does not compare other factors.  However, these are not important for this review.
* The test procedure described in the readme is more of an illustrative run - it is left to the user to interpret the results of the procedure to verify that the system is indeed running. As I understand it, the testing selects random excerpts from the example items as queries.  In my case, one of them selected from a long silent period at the end of one of the tracks, meaning that it did not match anything in the database.  A deterministic set of queries - perhaps even illustrating time-scale invariance - along with exact expected output would make verification of correct function easier.
