+++ resources/scripts/create_test_dataset/create_dataset.rb
@@ -6,6 +6,18 @@ require 'uri'
 require 'net/http'
 require 'fileutils'
 
+NUMBER_OF_FILES=5
+QUERY_LENGTH_IN_SEC = 15
+#Supply a random seed to make the procedure deterministic,
+#for random dataset supply RANDOM_SEED = Kernel.srand 
+RANDOM_SEED = 11
+
+QUERY_FOLDER = File.join('dataset','queries')
+DATASET_FOLDER = File.join('dataset','reference_items')
+SPEED_UP = true
+
+#seed the random number generator
+Kernel.srand(RANDOM_SEED)
 
 def file_length(file)
   $1.to_i if `sox "#{file}" -n stat 2>&1` =~ /.*Length .*:\s*(\d*)\..*/
@@ -14,7 +26,6 @@ end
 def seconds_to_samples(seconds)
   '%ds' % (seconds * 44100)
 end
-
 
 def cut_piece(input_file,target_file,start,piece_length)
   puts input_file
@@ -39,37 +50,63 @@ def compressor(input_file,target_file)
   system(command) 
 end
 
+def speed_change(input_file,target_file,percentage)
+  command = "sox \"#{input_file}\" \"#{target_file}\" speed #{percentage/100.0}"
+  puts command
+  system(command)
+end
 
-def download_audio_files(id_file,files) 	
-  ids = File.read(id_file).split("\n").shuffle[0..files]
+
+def download_audio_files(id_file,number_of_files,dataset_folder) 	
+  ids = File.read(id_file).split("\n").shuffle[0..(number_of_files -1)]
   ids.each do |id|
+
     url = "http://storage-new.newjamendo.com/download/track/#{id}/mp31/?from=app-00523d33"
 	  #url = url.gsub("mp31","mp32") #downloads better quality mp3
-	  output_file = "#{id}.mp3"
-    output_wav_file = "#{id}.wav"
-    command = "wget --no-check-certificate \"#{url}\" -O \"#{output_file}\""
+	  output_file = File.join(dataset_folder,"#{id}.mp3")
+    output_wav_file = File.join(dataset_folder,"#{id}.wav")
+    
     if File.exists?(output_wav_file)
-      puts "Skipping: #{output_file}"
+      puts "Skipping download: #{output_file}"
     else
-      system(command)
-      system("ffmpeg -i #{output_file} -ac 1 -ar 44100  -acodec  pcm_s16le #{output_wav_file}")
+      download_command = "wget -nv --no-check-certificate \"#{url}\" -O \"#{output_file}\""
+      system(download_command)
+      system("ffmpeg  -hide_banner -loglevel error -i #{output_file} -ac 1 -ar 44100  -acodec  pcm_s16le #{output_wav_file}")
     end
   end
 end
 
+FileUtils.mkdir_p(QUERY_FOLDER) unless File.exists? QUERY_FOLDER
+FileUtils.mkdir_p(DATASET_FOLDER) unless File.exists? DATASET_FOLDER
+#Download the audio files
+download_audio_files("jamendo_track_ids.csv",NUMBER_OF_FILES,DATASET_FOLDER)
 
-query_folder = 'query'
-dataset_folder = 'dataset'
+dataset_files = Dir.glob(File.join(DATASET_FOLDER,"*.wav"))
 
-Dir.mkdir(query_folder) unless File.exists? query_folder
-Dir.mkdir(dataset_folder) unless File.exists? dataset_folder
-files = 5
-download_audio_files("jamendo_track_ids.csv",files)
-Dir.glob("*.wav") do |input_file|
-  start = cut_random_piece(input_file,'temp.wav',15)
-  compressor('temp.wav',File.join(query_folder,start.to_s + '_' + input_file))
[…]
