### 2022-07-12 abb10f9
Removed non ascii char for compilation on windows
