+++ src/main/java/be/panako/strategy/panako/PanakoStrategy.java
@@ -224,7 +224,7 @@ public class PanakoStrategy extends Strategy {
 	private int mostCommonDeltaTforHitList(List<PanakoMatch> hitList) {
 		Map<Integer,Integer> countPerDiff = new HashMap<>();
 		hitList.forEach((hit)->{
-			int deltaT = hit.Δt();
+			int deltaT = hit.deltaT();
 			if(!countPerDiff.containsKey(deltaT))
 				countPerDiff.put(deltaT, 0);
 			countPerDiff.put(deltaT, countPerDiff.get(deltaT)+1);
@@ -337,7 +337,7 @@ public class PanakoStrategy extends Strategy {
 			 float frequencyFactor = 0;
 			 for(int i = 0 ; i < firstHits.size() ; i++) {
 				 PanakoMatch hit = firstHits.get(i);
-				 int diff = hit.Δt();
+				 int diff = hit.deltaT();
 				 if(diff == y1) {
 					 x1 = hit.queryTime;
 					 frequencyFactor = binToHz(hit.matchF1) / binToHz(hit.queryF1);
@@ -350,7 +350,7 @@ public class PanakoStrategy extends Strategy {
 			 float x2 = 0;
 			 for(int i = lastHits.size() - 1 ; i >= 0 ; i--) {
 				 PanakoMatch hit = lastHits.get(i);
-				 int diff = hit.Δt();
+				 int diff = hit.deltaT();
 				 if(diff == y2) {
 					 x2 = hit.queryTime;
 					 break;
@@ -372,7 +372,7 @@ public class PanakoStrategy extends Strategy {
 				 List<PanakoMatch> filteredHits = new ArrayList<>();
 				 
 				 hitlist.forEach( hit ->{				 
-					 float yActual = hit.Δt();
+					 float yActual = hit.deltaT();
 					 float x = hit.queryTime;
 					 float yPredicted = slope * x + offset;
 					 
