+++ src/main/java/be/panako/strategy/olaf/OlafMatch.java
@@ -56,11 +56,11 @@ public class OlafMatch {
 	public int queryTime;
 	
 	
-	public int Δt() {
+	public int deltaT() {
 		return matchTime - queryTime;
 	}
 	
 	public String toString() {
-		return String.format("%d %d %d %d", identifier, matchTime, queryTime, Δt());
+		return String.format("%d %d %d %d", identifier, matchTime, queryTime, deltaT());
 	}
 }
