### 2022-07-13 #4554 @JorenSix
@dpwe Thanks for the review, it raises a couple of issues that have been addressed and included in the paper (above) and in the readme. In the same order:

- I have made explicit how to contribute in the readme. See the [contribute](https://github.com/JorenSix/Panako#contribute) section.
- On being the only system against time-scale robustness: that claim indeed needs to be tapered. In the paper I have included a note on the matlab version of audfprint. Additionally i have changed ' [systems] lack robustness to speed changes' to 'robustness against significant speed changes of more than 5%'. The 'significant' also makes modifying a query to match a reference item unpractical. If you want to span a range of +-10% for speed-up, pitch-shift, time-stretch a lot of queries need to be generated. Perhaps uncommon - but Panako supports it - a pitch-shift of +5% can be combined with a time-stretch of -7% making the search space even larger.
- Thanks, I was indeed not aware of the Python version. A reference has been added to both the paper and the readme.
- I agree that Panako has its merits as a scalable acoustic fingerprinter, even without the time-scale invariance. Further evaluation is indeed always welcome. There is a paper submitted which evaluates Panako for a specific application with  low SNR. The application determines a great deal of these metrics (number of fingerprints per seconds, processing time, storage requirements, query performance, ...) . Evidently, detecting 'clean' duplicates is much less demanding than over-the-air, DJ-set song detection. Applications for speech have still other requirements.
- The test was indeed somewhat lacking. I have made the test procedure deterministic by providing a seed to the random number generator. Now it also includes the time-scale invariance: queries are sped up between 93% and 107%. Another thing that was missing was a proper explanation of the output of Panako. See the [Test and output](https://github.com/JorenSix/Panako#test) section of the readme for the change.
