+++ src/main/java/be/panako/strategy/olaf/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * The main algorithms for the Olaf Strategy: extracting fingerprints, indexing and querying.
+ */
+package be.panako.strategy.olaf;
