+++ src/main/java/be/panako/strategy/olaf/storage/OlafStorageKV.java
@@ -69,7 +69,7 @@ public class OlafStorageKV implements OlafStorage {
 	/**
 	 * A mutex for synchronization purposes
 	 */
-	private static Object mutex = new Object();
+	private static final Object  mutex = new Object();
 
 	/**
 	 * @return Returns or creates a storage instance. This should be a thread
