+++ src/main/java/be/panako/strategy/olaf/storage/OlafStorageMemory.java
@@ -52,7 +52,7 @@ public class OlafStorageMemory implements OlafStorage {
 	/**
 	 * A mutex for synchronization purposes
 	 */
-	private static Object mutex = new Object();
+	private static final Object mutex = new Object();
 
 	/**
 	 * @return Returns or creates a storage instance. This should be a thread
