+++ src/main/java/be/panako/strategy/panako/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Main algorithms for extracting, storing and querying fingerprints with the Panako strategy.
+ */
+package be.panako.strategy.panako;

[…]
