+++ .github/workflows/gradle.yml
@@ -4,8 +4,8 @@
 # documentation.
 # This workflow will build a Java project with Gradle and cache/restore any dependencies to improve the workflow execution time
 # For more information see: https://help.github.com/actions/language-and-framework-guides/building-and-testing-java-with-gradle
-
-name: Java CI with Gradle
+#https://github.com/JorenSix/Panako/actions/workflows/gradle.yml/badge.svg
+name: Build Panako with Gradle
 
 on:
   push:
