+++ .github/workflows/gradle.yml
@@ -23,6 +23,8 @@ jobs:
 
     steps:
     - uses: actions/checkout@v3
+    - uses: FedericoCarboni/setup-ffmpeg@v1
+      id: setup-ffmpeg
     - name: Set up JDK 11
       uses: actions/setup-java@v3
       with:
