### 2022-08-29 d1c595b
See issue 33, thanks @lucasso44
