### 2022-08-29 #4554 @JorenSix
Hi all,

I have been spending some time on improving Panako. Here is my answer to the review by @hendriks73

### Performance
The performance claims were indeed somewhat scattered and vague. A lot depends on use cases, reference database size and audio noise types but I have made a benchmark script to at least provide some reference points. The script doubles the reference database a couple of times and after each doubling two measurements are done:
- The time it took to double the database.
- The time it takes to run a number of queries.

The results show the behavior of the system for indexing and querying audio for a growing database.

https://github.com/JorenSix/Panako/blob/master/resources/scripts/benchmark/benchmark.textile

### Functionality documentation

I have improved the documentation significantly. Now every public member is documented with descriptions. All JavaDoc warnings have been fixed.

### Community guidelines

I have copied the community guidelines from the readme to the CONTRIBUTING.textile file.

### Tests

I have made a couple of unit tests to test the core functionality of the software. There is also a GitHub Action which builds the software and runs these test. Now every check in is ensured to at least compile and have that core functionality. The manual test scripts allow further functional verification.

### Other

* Panako is often used on servers and targets Java 11 LTS since Debian stable and other conservative server platforms, also still use JDK 11 as the default JDK. Of coarse it is no problem to run or build Panako with a newer JRE/JDK.
* Panako is currently not a JPMS module. The problem is that the dependencies (javalmdb, TarsosDSP, JGaborator) are not modularized as well. The long term plan is to modularize TarsosDSP and JGaborator myself (since I have control over those dependencies) and see if there is a replacement for javalmdb.
* Publishing a package is not a top priority as I do think Panako will be used mainly as an end-user application. However I am keeping GitHub Packages in mind as a possible solution.

### Typos

The typos have been fixed. Thanks!
