### 2022-09-06 2b42308
Meta data application and part list lengthconfiguration see #36
