+++ src/main/java/be/panako/strategy/olaf/storage/OlafStorageKV.java
@@ -40,7 +40,6 @@ import java.nio.ByteBuffer;
 import java.nio.charset.StandardCharsets;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
