### 2022-10-07 83bd3b9
Download does not follow redirects from http->https. Fix url to https
