### 2022-10-07 8020bd6
Extend monitor command to allow multiple files
