+++ src/main/java/be/panako/cli/Monitor.java
@@ -32,26 +32,28 @@
 *                                                                          *
 ****************************************************************************/
 
-
-
-
-
-
 package be.panako.cli;
 
 
+import java.io.File;
 import java.util.HashSet;
+import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.logging.Logger;
 
 import be.panako.strategy.QueryResult;
 import be.panako.strategy.QueryResultHandler;
 import be.panako.strategy.Strategy;
+import be.panako.util.Config;
+import be.panako.util.Key;
 import be.tarsos.dsp.util.AudioResourceUtils;
 
 /**
  * A command line application to monitor long audio files.
  */
 public class Monitor extends Application implements QueryResultHandler {
-
+	private final static Logger LOG = Logger.getLogger(Monitor.class.getName());
 	/**
 	 * Default constructor
 	 */
@@ -59,16 +61,80 @@ public class Monitor extends Application implements QueryResultHandler {
 
 	@Override
 	public void run(String... args) {
-		String inputResource = null;
-		if (args.length > 0){
-			inputResource = AudioResourceUtils.sanitizeResource(args[0]);
-		}else if (args.length == 0){
-			inputResource = Panako.DEFAULT_MICROPHONE;
-		}
 		Strategy strategy = Strategy.getInstance();
-		
+		int processors = availableProcessors();
+
+		if (args.length == 0){
+			String inputResource = Panako.DEFAULT_MICROPHONE;
+			Panako.printQueryResultHeader();
+			strategy.monitor(inputResource,1,new HashSet<Integer>(), this);
+			return;
+		}
+		// else get and process list of files
+
+		List<File> files = this.getFilesFromArguments(args);
+
+		if(files.size() > 1){
+			System.out.println("Processing " + files.size() + " queries on " + processors + " separate threads.");
+		}
+
 		Panako.printQueryResultHeader();
-		strategy.monitor(inputResource,1,new HashSet<Integer>(), this);
+
+		if(hasArgument("debug", args) || processors==1){
+			int taskNumber = 1;
+			for(File file: files){
+				new Monitor.MonitorTask(file.getPath(),taskNumber,files.size()).run();
+				taskNumber++;
+			}
+		}else{
+			ExecutorService executor = Executors.newFixedThreadPool(processors);
+			int taskNumber = 1;
+			for(File file: files){
+				executor.submit(new Monitor.MonitorTask(file.getPath(),taskNumber,files.size()));
+				taskNumber++;
+			}
+			executor.shutdown();
+			try {
+				//wait for tasks to finish
+				executor.awaitTermination(300, java.util.concurrent.TimeUnit.DAYS);
+				System.exit(0);
+			} catch (InterruptedException e1) {
+				//Thread was interrupted
+				LOG.severe("Did not finish all tasks, thread was interrupted!");
+			}
+		}
+	}
+
+	private static class MonitorTask implements Runnable, QueryResultHandler{
+		private final String path;
+		private final HashSet<Integer> emptyHashSet = new HashSet<Integer>();
+		private final Strategy strategy;
+		private final int numberOfQueryResults;
+		private final int taskNumber;
+		private final int totalTasks;
+
+		public MonitorTask(String path,int taskNumber, int totalTasks){
+			this.path = path;
+			this.numberOfQueryResults = Config.getInt(Key.NUMBER_OF_QUERY_RESULTS);
+			strategy = Strategy.getInstance();
+			this.taskNumber = taskNumber;
+			this.totalTasks = totalTasks;
+		}
+
+		@Override
[…]
