### 2022-10-08 #4554 @hendriks73
Apologies for not getting back to you earlier and thanks for the improvements, @JorenSix. From my POV, all boxes are checked now.

> The problem is that the dependencies (javalmdb, TarsosDSP, JGaborator) are not modularized as well. The long term plan is to modularize TarsosDSP and JGaborator myself (since I have control over those dependencies) and see if there is a replacement for javalmdb.

It's not pretty, but a JPMS module can be dependent on a plain old jar file (an [*automatic module*](https://stackoverflow.com/a/46742802/942774)).

> Publishing a package is not a top priority as I do think Panako will be used mainly as an end-user application. However I am keeping GitHub Packages in mind as a possible solution.

It's just something JOSS encourages and simplifies usage for people not wanting to `git clone` to get to your software. But I guess this is not crucial for acceptance.
