### 2022-10-10 #4554 @bmcfee
@JorenSix I gave the paper a read-through, and it mostly looks to be in good shape.  I did spot a couple of typos:

- line 7: "in large database" → "in a large database"
- line 64: "technoloy" → "technology"

I'd recommend hitting it with a spell checker just in case I missed anything.

The editorial bot gives a :+1: on DOIs (see above in thread). However, from reading the bibliography, I think there are a few references that are missing DOIs that should be corrected.  For example, all ISMIR publications are indexed on zenodo now and should have DOIs that you can find by following through from https://ismir.net/conferences/ .  Please double-check each of your citations for a DOI, and add them as needed.
