+++ paper.bib
@@ -46,6 +46,7 @@
   pages     = {7--13},
   title     = {{An Industrial-Strength Audio Search Algorithm}},
   year      = {2003},
+  doi       = {10.5281/zenodo.1416340}
 }
 
 @inproceedings{fenet2011pitch_shift_fingerprinting,
@@ -54,6 +55,7 @@
   pages     = {121--126},
   title     = {{A Scalable Audio Fingerprint Method with Robustness to Pitch-Shifting}},
   year      = {2011},
+  doi       = {10.5281/zenodo.1417593}
 }
 
 @article{holighaus2012framework,
@@ -83,20 +85,6 @@
 }
 
 
-@incollection{2013LesaffreIPEM,
-  author       = {{Lesaffre, Micheline}},
-  booktitle    = {{IPEM Institute for Psychoacoustics and Electronic Music : 50 years of electronic and electroacoustic music at the Ghent University}},
-  editor       = {{Jacobs, Greg}},
-  keywords     = {{psychoacoustics,electronic music,IPEM}},
-  language     = {{eng}},
-  pages        = {{23--25}},
-  publisher    = {{Metaphon}},
-  series       = {{Metaphon}},
-  title        = {{50 years of the Institute for psychoacoustics and electronic music}},
-  volume       = {{004}},
-  year         = {{2013}},
-}
-
 @article{cano2005fingerprinting_overview,
   author       = {Cano, Pedro and Batlle, Eloi and Kalker, Ton and Haitsma, Jaap},
   issn         = {0922-5773},
@@ -110,25 +98,12 @@
   year         = {2005},
 }
 
-
-@inproceedings{wang2003shazam,
-  author        = {Wang, Avery Li-Chun},
-  booktitle     = {Proceedings of the 4th {I}nternational {S}ymposium on {M}usic {I}nformation {R}etrieval ({ISMIR} 2003)},
-  location      = {Baltimore, MD, USA},
-  pages         = {7--13},
-  title         = {{An industrial-strength audio search algorithm}},
-  year          = {2003},
-}
-
 @inproceedings{fma2017fma,
   author        = {Michaël Defferrard and Kirell Benzi and Pierre Vandergheynst and Xavier Bresson},
   booktitle     = {Proceedings of the 18th {I}nternational {S}ymposium on {M}usic {I}nformation {R}etrieval ({ISMIR} 2017)},
   title         = {{FMA: a dataset for music analysis}},
   year          = {2017}
 }
-
-
-
 
 @article{brown1991calculation,
   title={Calculation of a constant Q spectral transform},
@@ -259,6 +234,7 @@ booktitle	= {ISMIR 2020 - 21st International Society for Music Information Retri
   pages         = {1--6},
   title         = {{Panako - A scalable acoustic fingerprinting system handling time-scale and pitch modification}},
   year          = {2014},
+  doi           = {10.5281/zenodo.1416190}
 }
 
 @article{six2015synchronizing,
@@ -282,10 +258,11 @@ booktitle	= {ISMIR 2020 - 21st International Society for Music Information Retri
 }
 
 @inproceedings {black2018unmixdb,
-  title = {Unmixdb: Een dataset voor het ophalen van DJ-mixinformatie},
+  title = {Unmixdb: A Dataset for DJ-Mix Information Retrieval},
   author = {Schwarz, Diemo and Fourer, Dominique},
   boektitel = {19th International Symposium on Music Information Retrieval (ISMIR)},
-  year = {2018}
+  year = {2018},
+  doi = {10.5281/zenodo.1422385}
 }
 
 @dataset{schwarz_diemo_2018_1422385,
[…]
