### 2022-10-10 #4554 @JorenSix
Hi @bmcfee,

Thanks for the check and the recommendations. I did not know about the ISMIR/zenodo DOI's. They have been added.

Also the typos have been fixed (found one more mistake).
