### 2022-10-10 #4554 @bmcfee
Thanks @JorenSix - it looks like there are still a few missing DOIs, both from ISMIR and other venues.  The Cano et al. paper has a DOI https://doi.org/10.1007/s11265-005-4151-3 ; Sonnleitner, Artz & Widmer (2011) also should have one.  The ISMIR LBDs (AFAIK) do not receive DOI's, but they should be cited consistently and with proper formatting.

Please give each bibliography entry a thorough check before we move this on to the next stage.
