+++ paper.bib
@@ -14,10 +14,10 @@
   year={2011}
 }
 
-@article{velasco2011constructing,
+@inproceedings{velasco2011constructing,
   title={Constructing an invertible constant-Q transform with non-stationary Gabor frames},
   author={{Velasco, Gino Angelo and Holighaus, Nicki and D{\"o}rfler, Monika and Grill, Thomas}},
-  journal={Proceedings of DAFX11, Paris},
+  booktitle={{Proceedings of the 14th International Conference on Digital Audio Effects (DAFx-11)}},
   volume={33},
   year={2011}
 }
@@ -96,13 +96,15 @@
   title        = {A Review of Audio Fingerprinting},
   volume       = {41},
   year         = {2005},
+  doi          = {10.1007/s11265-005-4151-3}
 }
 
 @inproceedings{fma2017fma,
   author        = {Michaël Defferrard and Kirell Benzi and Pierre Vandergheynst and Xavier Bresson},
   booktitle     = {Proceedings of the 18th {I}nternational {S}ymposium on {M}usic {I}nformation {R}etrieval ({ISMIR} 2017)},
   title         = {{FMA: a dataset for music analysis}},
-  year          = {2017}
+  year          = {2017},
+  doi           = {10.5281/zenodo.1414728}
 }
 
 @article{brown1991calculation,
@@ -113,7 +115,8 @@
   number={1},
   pages={425--434},
   year={1991},
-  publisher={Acoustical Society of America}
+  publisher={Acoustical Society of America},
+  doi = {10.1121/1.400476}
 }
 
 @article{brown1992efficient,
@@ -124,7 +127,8 @@
   number={5},
   pages={2698--2701},
   year={1992},
-  publisher={Acoustical Society of America}
+  publisher={Acoustical Society of America},
+  doi={10.1121/1.404385}
 }
 
 @article{comer1979ubiquitous,
@@ -140,7 +144,7 @@
 
 @INPROCEEDINGS{9044274,
   author={Mehmood, Zahid and Ashfaq Qazi, Khurram and Tahir, Muhammad and Muhammad Yousaf, Rehan and Sardaraz, Muhammad},
-  booktitle={2020 6th Conference on Data Science and Machine Learning Applications (CDMA)}, 
+  booktitle={6th Conference on Data Science and Machine Learning Applications (CDMA 2020)},
   title={Potential Barriers to Music Fingerprinting Algorithms in the Presence of Background Noise}, 
   year={2020},
   volume={},
@@ -168,13 +172,6 @@
  keywords = {Applying MIR, collaboration, digital music archives, documentation},
 }
 
-@inproceedings{gfeller2020denoise,
-title	= {Learning to Denoise Historical Music},
-author	= {Beat Gfeller and Dominik Roblek and Marco Tagliasacchi and Pen Li},
-year	= {2020},
-booktitle	= {ISMIR 2020 - 21st International Society for Music Information Retrieval Conference}
-}
-
 
 @inproceedings{sonnleitner2014quad_based_fingerprinter,
   author    = {Reinhard Sonnleitner and Gerhard Widmer},
@@ -182,10 +179,11 @@ booktitle	= {ISMIR 2020 - 21st International Society for Music Information Retri
   title     = {{Quad-based Audio Fingerprinting Robust To Time And Frequency Scaling}},
   year      = {2014},
 }
+
 @inproceedings{ellis20142014,
   title={The 2014 labrosa audio fingerprint system},
   author={Ellis, Daniel},
-  booktitle={ISMIR},
+  booktitle={ISMIR 2014 MIREX abstract},
   year={2014}
 }
 
@@ -216,11 +214,13 @@ booktitle	= {ISMIR 2020 - 21st International Society for Music Information Retri
[…]
