+++ paper.bib
@@ -220,7 +220,7 @@
 
 @inproceedings{six2021panakovtwo,
   author       = {{Six, Joren}},
-  booktitle    = {{Extended abstracts for the Late-Breaking Demo Session of the 22st International Society for Music Information Conference (ISMIR 2021) }}
+  booktitle    = {{Extended abstracts for the Late-Breaking Demo Session of the 22st International Society for Music Information Conference (ISMIR 2021) }},
   keywords     = {{Acoustic fingerprinting}},
   language     = {{eng}},
   location     = {{Online}},
@@ -290,7 +290,7 @@
 @inproceedings {kim2020djset,
   title = {A Computational Analysis of Real-World DJ Mixes using Mix-To-Track Subsequence Alignment},
   author = {Taejun Kim and Minsuk Choi and Evan Sacks and Yi-Hsuan Yang and Juhan Nam},
-  boektitel = {21st International Society for Music Information Retrieval Conference (ISMIR), 2020},
+  bootitle = {Proceedings of the 21st International Society for Music Information Retrieval Conference (ISMIR), 2020},
   year = {2020},
   doi = {10.5281/zenodo.4245544}
 }
