### 2022-10-11 ecd9f1b
bibtex formatting
