+++ paper.bib
@@ -183,7 +183,7 @@
 @inproceedings{ellis20142014,
   title={The 2014 labrosa audio fingerprint system},
   author={Ellis, Daniel},
-  booktitle={ISMIR 2014 MIREX abstract},
+  booktitle={{MIREX abstracts of the 15th International Symposium on Music Information Retrieval (ISMIR 2014)}},
   year={2014}
 }
 
@@ -212,7 +212,7 @@
 @inproceedings{sonnleitner2016landmark,
   title={Landmark-based audio fingerprinting for DJ mix monitoring.},
   author={Sonnleitner, Reinhard and Arzt, Andreas and Widmer, Gerhard},
-  booktitle={ISMIR},
+  booktitle={{Proceedings of the 17th International Symposium on Music Information Retrieval (ISMIR 2016)}},
   pages={185--191},
   year={2016},
   doi={10.5281/zenodo.1417008}
@@ -260,7 +260,7 @@
 @inproceedings {black2018unmixdb,
   title = {Unmixdb: A Dataset for DJ-Mix Information Retrieval},
   author = {Schwarz, Diemo and Fourer, Dominique},
-  booktitle = {{Proceedings of the 19th International Symposium on Music Information Retrieval (ISMIR)}},
+  booktitle = {{Proceedings of the 19th International Symposium on Music Information Retrieval (ISMIR 2018)}},
   year = {2018},
   doi = {10.5281/zenodo.1422385}
 }
