### 2022-10-11 #4554 @JorenSix
@bmcfee Thanks again for the check. After a couple of tries the formatting of different ISMIR papers/abstracts is consistent now and additional DOI's have been added. I did not find DOI's for DAFx and AES conference papers.
