+++ paper.bib
@@ -5,15 +5,6 @@
   title = {{A Robust Audio Fingerprinter Based on Pitch Class Histograms - Applications for Ethnic Music Archives}},
   year = {2012}}
 
-
-@article{velasco2011constructing,
-  title={Constructing an invertible constant-Q transform with non-stationary Gabor frames},
-  author={{Velasco, Gino Angelo and Holighaus, Nicki and D{\"o}rfler, Monika and Grill, Thomas}},
-  journal={Proceedings of DAFX11, Paris},
-  volume={33},
-  year={2011}
-}
-
 @inproceedings{velasco2011constructing,
   title={Constructing an invertible constant-Q transform with non-stationary Gabor frames},
   author={{Velasco, Gino Angelo and Holighaus, Nicki and D{\"o}rfler, Monika and Grill, Thomas}},
