### 2022-10-11 #4554 @JorenSix
Ok, I still had a duplicate entry in the paper.bib (explaining the error in the check references I think): a change from an incorrect @article to a correct @inproceedings. Now fixed.
