### 2022-10-11 c2cd1cf
Caching improvements: to avoid repeated feature extraction caching is now implemented with a content based hash and not on a path of a file.
