### 2022-10-11 #4554 @JorenSix
@bmcfee Thanks for the quick follow-up. I have tagged [a release on Github](https://github.com/JorenSix/Panako/releases/tag/joss) and archived it on Zenodo with the DOI [10.5281/zenodo.7185611](https://doi.org/10.5281/zenodo.7185611)
